move duplicate do_not_go_off_screen into common helper module
