+++ angry_plant.rb
@@ -1,5 +1,7 @@
+require './game_object_helpers'
 class AngryPlant
-attr_reader :angry_x, :angry_y, :center_x, :center_y, :radius # :ishappy
+  attr_reader :angry_x, :angry_y, :center_x, :center_y, :radius # :ishappy
+  include GameObjectHelpers
 
   def initialize
     @frame_count = 0
@@ -8,7 +10,7 @@ attr_reader :angry_x, :angry_y, :center_x, :center_y, :radius # :ishappy
     @z = 1
     @scale = 0.1
     @AngryPlant1 = Gosu::Image.new("images/angryplant_1.png")
-    @AngryPlant2 =Gosu::Image.new("images/angryplant_2.png")
+    @AngryPlant2 = Gosu::Image.new("images/angryplant_2.png")
     @happy_image = Gosu::Image.new("images/happyplant.png")
     @height = @AngryPlant2.height
     @width = @AngryPlant2.width
@@ -36,7 +38,7 @@ attr_reader :angry_x, :angry_y, :center_x, :center_y, :radius # :ishappy
       @y += rand(-5..5)
     end
     @frame_count += 1
-    
+
     recalc_center
     @angry_x = @x
     @angry_y = @y
@@ -45,21 +47,6 @@ attr_reader :angry_x, :angry_y, :center_x, :center_y, :radius # :ishappy
 
   def make_happy
     @is_happy = true
-  end
-
-  def do_not_go_off_screen
-    if @x < 0
-      @x = 0
-    end
-    if @y < 0
-      @y = 0
-    end
-    if @x > MyWindow::WIDTH - (@width * @scale)
-      @x = MyWindow::WIDTH - (@width * @scale)
-    end
-    if @y > MyWindow::HEIGHT - (@height * @scale)
-      @y = MyWindow::HEIGHT - (@height * @scale)
-    end
   end
 
   def recalc_center
