+++ water.rb
@@ -1,5 +1,8 @@
+require './game_object_helpers'
+
 class Water
   attr_reader :center_x, :center_y, :radius
+  include GameObjectHelpers
 
   def initialize
     @x = 0
@@ -43,21 +46,6 @@ class Water
     end
     recalc_center
     do_not_go_off_screen
-  end
-
-  def do_not_go_off_screen
-    if @x < 0
-      @x = 0
-    end
-    if @y < 0
-      @y = 0
-    end
-    if @x > MyWindow::WIDTH - @width
-      @x = MyWindow::WIDTH - @width
-    end
-    if @y > MyWindow::HEIGHT - @height
-      @y = MyWindow::HEIGHT - @height
-    end
   end
 
   def recalc_center
